refactor(ios): extract TodaySections + DailyBriefing cerulean header
D.5 — DailyBriefing header label: white/60 → cerulean/60. The card
already wore the cerulean rim that signals a Brett AI surface, but the
"DAILY BRIEFING" label was a neutral white. Aligning with the
DESIGN_GUIDE's rule that AI surfaces use cerulean accents throughout —
and matches the desktop's DailyBriefing header treatment.

D.6 — TodayPage bucketing split:
- Move `TodaySections` + `TodaySections.bucket` to a new
  `Views/Today/TodaySections.swift`. Pure value type, no SwiftUI,
  no @MainActor isolation — can be unit-tested with fixture Items
  without rendering a view.
- BrettTests/Views/TodaySectionsTests.swift covers the rule set:
  yesterday → overdue, today → today, completed-today → doneToday,
  completed-yesterday → dropped entirely, archived → dropped,
  pendingDoneID holds the item in its prior section, no-due-date
  items are ignored, and sorting puts earliest-due first.

TodayPage.swift gets ~90 lines lighter as a side effect. Same
behaviour as before — the extraction is pure textual.
